Added functionality and bug-fix
Added some extra functionality to the Tools source file so that specific
characters may be removed from a string object. This is handy and can be
used in conjunction with the regular text parser. The Animation class
had a small bug in the timer that needed attention. The Animation class
can now also return a flag when the animation is done rendering, so that
the game object that contains the animation component knows when to
destroy itself.