Mouse Collision & Updated Sprite Objects
Mouse collision is now available in the InputManager, so that buttons
may be hovered over or game objects may be clicked on. The Sprite class
and its sub-classes Animation and Text have undergone a few small
updates including an updated inner loop that is and looks cleaner, the
Text object not needing to use the modelview matrix directly - text is
now rendered a little differently. The Animation class also underwent a
few small tweaks.