Updated Shaders
Small tweaks to how the main vert and frag shaders are written. There is
now a flag in the frag shader to determine if the object should be
textured or not. Previously there were 2 separate debug shaders (now
removed) that were used to render un-textured objects and then there
were 2 main shaders to render as normal with textures. What a pilava -
all better now!