Big Changes
The game state machine received an overhaul to represent a more robust system. The GameObject class now has support for a texture matrix and normal matrix so all objects in the scene can have their lighting and texturing properly manipulated. The DebugManager has a better way to manage its buffers.