Big Changes
A few cosmetic changes here and there in terms of code. The ShaderManager is now known as the PipelineManager and controls everything from the shaders to the entire graphics pipeline and how data is passed through it. The game states needed an update to reflect the new state machine. Sprite and Text are also different as they handle the way they render their "quads" in a different manner. Finally, no more TextureManager - one less Singleton!! All textures are managed via the Texture class.